web-scraper-python-main/backend/requirements.txt
flask
requests
beautifulsoup4
youtube-transcript-api
duckduckgo-search
google-generativeai
python-dotenv
flask-cors